Update patch set 1
Patch Set 1:

Liang, looks good, the comment at the end on the python library.  the LOG containers and logging in general is part of the OOM deployment and will benefit from standard Kubernetes resiliency and scaling for performance on both the emission and capture side.  Log retention will be configured in the undercloud that Kubernetes runs on

Patch-set: 1