Update patch set 1
Patch Set 1:

(1 comment)

Only 2 hours 33 minutes was given for review and comments after the patch was submitted. Then it was quickly merged. It will be much nicer to give more time for review and comment. IMHO, this API design is wrong and doesn't meet OOF requirement, though it was super quickly merged. More specific comments are in line.

Patch-set: 1