Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Thank you Hong Hui Xiao for adding clarification.

Sorry I still have to give -1 because I need to wait for Xinhui's guidance on this. Please see my specific comment to Xinhui. If Xinhu is satisfied with this clarification that should be used to focus on the scope of spec with use cases, I will be fine.

Please understand that your spec is great. I just need to wait for PTL's explicit guidance which can be applied to all specs.

Thank you.

Patch-set: 6
Label: Code-Review=-1