Update patch set 6
Patch Set 6:

It doesn't matter whether it is an MVP feature or not, and doesn't matter what Architecture Committee says. In fact, MVP feature should have higher bar to pass because it is "MVP".

There is a comment. There is a response to the comment.

Please explicitly let us know if this response satisfactorily addressed the comment, or not.

Patch-set: 6