Update patch set 5
Patch Set 5: Code-Review-2

The handling of patch 30027 "spec for Container based network service/function" has set up a precedent of how we need to handle our features in a fair and consistent way in our project.

I understand the rationale of handling of 30027 is that:

- The technical spec needs to be approved first
- If there are still comments on the technical specs, creation of repo (if needed) cannot be approved. E.g. 30027 and its repo creation request
- Logically, it implies that if there are still comments on the technical specs, we need to hold on pushing the code (because even repo cannot be approved).

Based on above rationale, the same principle will be applied to all of our features in a fair and consistent way.

Thus I give -2 because the specs 28973 is not finished and approved yet. The code review should be postponed after we finish and approve the specs.

Patch-set: 5