Update patch set 3
Patch Set 3: Code-Review-2

Many thanks to Honghui for his dedication and hard work.

The following are my comments. Please understand that the purpose is for our common goal of a stable platform (S3P), and to make a better quality of Beijing (R2) Release and future releases.

While I understand the intention is the narrowed scope of API definition, the full and normative definition of API and resources will not be considered (Line 56-58), I really worried about its impact on overall platform stability and scalability.

For example, current definition is only limited to "host" and related "hypervisor" resources, the questions are:

* how can we ensure that when we attempt to scale to other resources, not only the atomic resources but also composite resources?
* Do we have to design another spec?
* How to ensure the new specs of bigger scope is backward compatible with this spec?
* How much extra work is needed to maintain the backward compatibility?
* Your parser and validator etc. will certainly be impacted, and the numerous patches of code in long term may become much more complex than it should be if we design it well at the very beginning.

Without complete understanding of above questions, platform stability and scalability, and even the performance will be adversary impacted in the future.

Given those unknowns, I really worried.

I suggest that we de-scope it from Beijing Release (R2) in order not to have an adversary impact on not only the quality of Beijing Release but also the S3P for the future.

Instead, let us focus on the the design of a full, normative definition of API and resources in the bigger picture. Given the time constraints, the design work can be targeted for Casablanca which is more reasonable.

Once we can ratify the full API design, we can design a fully scalable YAML parser and API generator. The last step is coding. Comprehensive testing is also a necessary step.

In conclusion, let us de-scope it from Beijing, focus on full, normative API design for Casablanca, and a PoC after all design specs are agreed.

Thus I gave -2.

Patch-set: 3