Update patch set 7
Patch Set 7:

> Many thanks to Honghui for his dedication and hard work.
 > 
 > The following are my comments. Please understand that the purpose
 > is for our common goal of a stable platform (S3P), and to make a
 > better quality of Beijing (R2) Release.
 > 
 > While this feature and spec talks about specific technique (Pecan)
 > for performance improvement, which sounds very good, I really
 > worried about how it impacts overall S3P goal, and how it improves
 > the quality R2 deliverable if we view it in a bigger picture.
 > 
 > S3P means Stability, Scalability, Security and Performance. It is
 > the goal of Beijing Release (R2). However, this spec seems to
 > address a small portion of Performance aspects. Thus the specific
 > questions are:
 > - which performance metrics and benchmarks are targeted?
 > - which use cases are targeted that meets above performance metrics
 > and benchmarks?
 > - which conditions are prerequisites to achieve the targeted
 > performance metrics?
 > - How do they compare to benchmarks?
 > - When we consider performance improvement, have we considered its
 > other impacts, especially adversary impacts on other platform
 > metrics such as:
 > * How does it impact or not impact platform stability?
 > * How does it impact or not impact platform scalability?
 > * How does it impact or not impact platform security?
 > 
 > Given those unknowns, I really worried.
 > 
 > I suggest that we de-scope it from Beijing Release (R2) in order
 > not to have an adversary impact on a quality Beijing Release.
 > Instead, it should start with a PoC, understand all those metrics,
 > benchmarks and potential impacts, test with real use cases, collect
 > all relevant test data regarding S3P metrics, establish a data
 > model, normalize the data collection and analyze the data so that
 > we can completely understand its benefits and potential
 > positive/negative impacts on the platform.
 > 
 > Thus I gave -2.

Hi Bin.


First, it is not Pecan which is the specific technology. It is mainly eventlet+wsgi that bump up the performance. 


And, as we previously discussed, S3P improvement is too big for one spec. So, I think it is good to just focus on Performance aspect, as long as there is no negative impact on other aspects, right?


For all these questions, I should already have described them in the spec. For example, performance metric is about improve the speed of handling api request.  If you think there is not properly described, I guess the better way is add comment inline, and I will re-word it. If you find there is any missing description, I will add them to spec.


And lastly,  this spec will NOT impact what we already have in multi-cloud(also mentioned in the spec), because of the way how we plan to implement it. I see no reason to delay or de-scope it in B release.

Patch-set: 7
Label: Code-Review=0