Update patch set 3
Patch Set 3:

Hi Bin.


I understand your concern. So the spec proposed add code in a separate path, which means the existing multicloud platform will not be affected at all. 


And I really don't want to mess this spec with "full normative api definition". Because "full normative api definition" is more like a model definition to me, and it is story. This spec is just to improve how we expose api in multicloud. I guess I have made that clear in my spec, but if not, pls let me know, I will re-word the spec.


Backward compatibility.  Since there is no such plan for "full normative api definition", I don't see backward in the picture.  And if the potential future work wants to reuse the code of this spec[1-2], you may find there is just about 2 hundreds of code. I guess it is easy for any programmer to understand and reuse them. If it is not, the discussion should be in the code itself, not in spec.


I know model driven design is great, but that is a huge work. And it is totally different from the purpose of this spec. What this spec trys to achieve, is to improve multicloud in small step. 


[1] https://gerrit.onap.org/r/#/c/35957/
[2] https://gerrit.onap.org/r/#/c/35637/

Patch-set: 3