Update patch set 7
Patch Set 7:

Honghui,

I am really sorry for your feeling of work being delayed. Although the feeling may not be correct, the most important is to do things right.

Because of fundamental issue in the methodology, it is very challenging to comment inline. You may see why as I comment more as follows.

(1) What is the best technical option to improve performance

Thank you for clarifying that it is eventlet+wsgi.

In the spec, there are 2 options, one is current django and the other one is eventlet+pecan (line 60). on Jan 28, there was a comment that "nginx + uwsgi to improve the performance by supporting concurrent API processing. We don't need another framework to address this issue.". And your response on Jan 30 was that "it might not be ideal api framework. I will add more details in new version, we can carry on discussion in new version."

Maybe I misunderstood. But I was expecting a comparison between nginx+uwsgi and eventlet+pecan, so that we can know which one is the best choice technically, and what are the impacts on current framework and amount of work for framework evolution.

However, entire Option 1 (Line 29-54) is talking about current django framework, and its cons. Option 2 is talking about eventlet+pecan. I didn't see any other options/solutions be discussed, analyzed, compared before we can definitively say let's go with eventlet+pecan.

If there is indeed no other options, it is good. But there was proposal about the option of nginx+uwsgi, and django+celery, I am just curious why those 2 options are ignored? If eventlet+pecan is indeed the best option, can we list all of those options, give a deep dive of technical pros and cons, and analyze what is the most applicable to our framework?

I am indeed not convinced why your solution is the best and most applicable to our framework.

(2) The implementation of your proposal is not mature for R2

On Jan 24, yourself already recognized that "And I don't think this api framework will be mature in R2." when you responded to another comment.

I understand that it is configurable option, and secondary to current framework. However, there is still risk of putting an immature and unstable solution into Beijing release, which is targeted for S3P - Stability is the most important.

Thus I am wondering what is the rationale here? Why cannot we start with a PoC, and put it into release after it is stable and mature? If we know it is unstable and immature, why cannot we wait until it is stable and mature?

Patch-set: 7