Update patch set 3
Patch Set 3:

(1) Review history in this thread may give everyone a clearer picture of discussion context.

(2) It is not like "go back to the API design". I see it as "API design has never been mature or ready".

(3) We are still discussing the API design, and now the question is what this design is for? Or what is the scope of this design? So we need your confirmation of Honghui's clarification.

Patch-set: 3