Update patch set 1
Patch Set 1:

It is great to have dedicate spec to design the image service. So would you please elaborate the consumers of this image service, and design the interaction and APIs between consumers and multicloud servcies(detail to framework and plugins which might be concerned)?
And we are approaching code freeze quickly, so I wonder if this feature be still part of Beijing Release? 
Thanks

Patch-set: 1