Update patch set 3
Patch Set 3:

Hi Bin Sun, as we discussed, please realize that not all VIM/Cloud's oslo events can be listened. So the design/implementation should allow multicloud plugin to feed the events to this "event federation" service.
It will be great to have some sequence diagram to elaborate how it can be achieved with your design. 
Thanks.

Patch-set: 3
Label: Code-Review=0