Create definitions for multiple vf-modules
This fix does not change distribution when there is one
dummy heat and one cloud technology artifact.

When there is more than one dummy heat or more than one
cloud technology artifact forwarder will try to match them
base on the name of the heat file that should match the prefix
of the name of cloud artifact. There is no need to change use
cases of vFW and EdgeX showcased for Dublin and ElAlto neither
their descriptions should be changed.

Issue-ID: MULTICLOUD-869
Signed-off-by: Lukasz Rajewski <lukasz.rajewski@orange.com>
Change-Id: I4d44ea22b412546806ee7dbcdb7e3c9ed644349a