Fixed distribution of k8s Helm Artifact
Cloud Technology Artifacts before were belonging
to each vf module. Now they are only on the list of
Cliud Archive artifacts. In coinsequence, in order
to do the mapping of CTA to vf-module we need to
look in the overall artifacts list. The algorithm
is the same so nothing should be changed except the
list we search in.

Issue-ID: SDC-2972
Signed-off-by: Lukasz Rajewski <lukasz.rajewski@orange.com>
Change-Id: I7336b494bfaf0ffb7992cf779391abaf16274b5b