Update patch set 1
Patch Set 1:

(1 comment)

just an admin question: shall we note indicate that it is for the maintenance release not the release? 
the tags on honolulu have been set on OOM so I am not sure this patch can be part of the honolulu release but for sure it is on the OOM honolulu branch that will be retagged for the maintenance release.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>